Refactor: move theme to separate file.
